requirements.txt
# ==========================================
# SANCHALAK AI - Complete Requirements
# ==========================================
# Python 3.10+ required
# Last verified: January 2026

# -----------------------------------------
# Core Web Framework (Backend)
# -----------------------------------------
fastapi>=0.109.0,<0.116.0
uvicorn[standard]>=0.27.0,<0.35.0
python-multipart>=0.0.6,<0.1.0
starlette>=0.35.0,<0.46.0

# -----------------------------------------
# Security & Authentication
# -----------------------------------------
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0

# -----------------------------------------
# Data Validation
# -----------------------------------------
pydantic>=2.5.3,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# -----------------------------------------
# Data Processing
# -----------------------------------------
pandas>=2.1.4,<2.3.0
numpy>=1.26.3,<2.0.0
polars>=0.20.0,<2.0.0
openpyxl>=3.1.0,<4.0.0
xlrd>=2.0.0,<3.0.0
pyarrow>=14.0.1,<15.0.0

# -----------------------------------------
# Database (REQUIRED)
# -----------------------------------------
sqlalchemy>=2.0.25,<3.0.0
psycopg2-binary>=2.9.9,<3.0.0
alembic>=1.13.0,<2.0.0

# -----------------------------------------
# Machine Learning - Core
# -----------------------------------------
scikit-learn>=1.3.0,<1.7.0
joblib>=1.3.0,<2.0.0
xgboost>=2.0.0,<3.0.0
lightgbm>=4.0.0,<5.0.0
shap>=0.44.0,<1.0.0
scipy>=1.11.0,<2.0.0


# -----------------------------------------
# ML Experiment Tracking
# -----------------------------------------
mlflow>=2.9.0,<3.0.0

# -----------------------------------------
# Report Generation & Export
# -----------------------------------------
Jinja2>=3.1.2,<4.0.0
weasyprint>=60.1,<61.0
python-docx>=1.1.0,<2.0.0
python-pptx>=0.6.22,<1.0.0

# -----------------------------------------
# Knowledge Graph
# -----------------------------------------
neo4j>=5.14.1,<6.0.0

# -----------------------------------------
# Data Streaming
# -----------------------------------------
kafka-python>=2.0.2,<3.0.0

# -----------------------------------------
# Vector Store & Embeddings
# -----------------------------------------
qdrant-client>=1.7.0,<2.0.0
fastembed>=0.1.0,<1.0.0
chromadb>=0.4.0,<1.0.0
sentence-transformers>=2.2.0,<3.0.0

# -----------------------------------------
# Visualization
# -----------------------------------------
plotly>=5.18.0,<6.0.0
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0
networkx>=3.0,<4.0.0

# -----------------------------------------
# LLM Integration
# -----------------------------------------
openai>=1.12.0,<2.0.0
google-genai>=1.0.0

# -----------------------------------------
# HTTP Clients
# -----------------------------------------
requests>=2.31.0,<3.0.0
httpx>=0.25.0,<1.0.0

# -----------------------------------------
# Utilities
# -----------------------------------------
python-dateutil>=2.8.2,<3.0.0
python-dotenv>=1.0.0,<2.0.0
typing_extensions>=4.9.0,<5.0.0

# -----------------------------------------
# CLI & Progress Bars
# -----------------------------------------
tqdm>=4.65.0,<5.0.0
colorama>=0.4.6,<1.0.0

# -----------------------------------------
# Logging
# -----------------------------------------
loguru>=0.7.0,<1.0.0
rich>=13.0.0,<15.0.0

# -----------------------------------------
# Testing (Optional but recommended)
# -----------------------------------------
pytest>=7.0.0,<9.0.0
